requirements.txt
torch
torchvision
torchaudio
tensorflow
opencv-python
pillow
numpy
matplotlib
scikit-learn
tqdm
fastapi
uvicorn
python-multipart
